Add tables and wrapline scripts
The tables.vim script actually duplicates some of the functionality of
the existing CSV script. I didn't realize when I started it that
something like this already existed. One difference with this script is
that tables can have blank cells. I'm not sure if the CSV plugin has
this, but this script also has an "undo" function (UndoTable()) for
turning the table back into an ascii table with pipe characters.

The Table() function can't handle tables on the last line of the file.

The WrapLine() function wraps the current line into multiple lines of
the specified length with the default being 70 bytes. Characters with
multi-byte representations are counted by byte, not character. This is
confusing because the statusline also counts by byte index, but 70
character lines without multi-byte characters will look longer.